Fix for settind download priorities with webserver
Apparently, too much type conversion caused the bug. (?)

+++ trunk/src/WebServer.h
@@ -139,7 +139,7 @@ class DownloadFiles {
 		long		lTransferringSourceCount;
 		long		lSourceCountA4AF;
 		double		fCompleted;
-		long		lFilePrio;
+		uint32		lFilePrio;
 		bool		bFileAutoPriority;
 		wxString	sFileHash;
 		wxString	sED2kLink;
